+++ setup.py
@@ -42,9 +42,8 @@ setuptools.setup(
     'console_scripts': [
         'get_promice_data = pypromice.get.get_promice_data:get_promice_data',
         'get_l0tx = pypromice.tx.get_l0tx:get_l0tx',
-        'get_l2 = pypromice.process.get_l2:get_l2',
-        'join_l2 = pypromice.process.join_l2:join_l2',
         'get_l3 = pypromice.process.get_l3:get_l3',
+        'join_l3 = pypromice.process.join_l3:join_l3',
         'get_watsontx = pypromice.tx.get_watsontx:get_watsontx',
         'get_bufr = pypromice.postprocess.get_bufr:main',
         'get_msg = pypromice.tx.get_msg:get_msg'
