+++ src/pypromice/process/L1toL2.py
@@ -30,18 +30,7 @@ def toL2(
     eps_clear=9.36508e-6,
     emissivity=0.97,
 ) -> xr.Dataset:
-    '''Process one Level 1 (L1) product to Level 2.
-    In this step we do:
-        - manual flagging and adjustments
-        - automated QC: persistence, percentile
-        - custom filter: gps_alt filter, NaN t_rad removed from dlr & ulr
-        - smoothing of tilt and rot
-        - calculation of rh with regards to ice in subfreezin conditions
-        - caluclation of cloud coverage
-        - correction of dsr and usr for tilt
-        - filtering of dsr based on a theoritical TOA irradiance and grazing light
-        - calculation of albedo
-        - caluclation of directional wind speed
+    '''Process one Level 1 (L1) product to Level 2
 
     Parameters
     ----------
@@ -96,20 +85,10 @@ def toL2(
     ds['dlr'] = ds.dlr.where(ds.t_rad.notnull())
     ds['ulr'] = ds.ulr.where(ds.t_rad.notnull())
 
-    # calculating realtive humidity with regard to ice
     T_100 = _getTempK(T_0)
     ds['rh_u_cor'] = correctHumidity(ds['rh_u'], ds['t_u'],
                                      T_0, T_100, ews, ei0)
 
-    if ds.attrs['number_of_booms']==2:
-        ds['rh_l_cor'] = correctHumidity(ds['rh_l'], ds['t_l'],
-                                         T_0, T_100, ews, ei0)
-
-    if hasattr(ds,'t_i'):
-        if ~ds['t_i'].isnull().all():
-            ds['rh_i_cor'] = correctHumidity(ds['rh_i'], ds['t_i'],
-                                             T_0, T_100, ews, ei0)
-    
     # Determiune cloud cover for on-ice stations
     cc = calcCloudCoverage(ds['t_u'], T_0, eps_overcast, eps_clear,        # Calculate cloud coverage
                            ds['dlr'], ds.attrs['station_id'])
@@ -197,50 +176,20 @@ def toL2(
         ds['precip_u_cor'], ds['precip_u_rate'] = correctPrecip(ds['precip_u'],
                                                                 ds['wspd_u'])
     if ds.attrs['number_of_booms']==2:
+        ds['rh_l_cor'] = correctHumidity(ds['rh_l'], ds['t_l'],           # Correct relative humidity
+                                         T_0, T_100, ews, ei0)
+
         if ~ds['precip_l'].isnull().all() and precip_flag:                     # Correct precipitation
             ds['precip_l_cor'], ds['precip_l_rate']= correctPrecip(ds['precip_l'],
                                                                    ds['wspd_l'])
 
-    # Get directional wind speed 
-    ds['wdir_u'] = ds['wdir_u'].where(ds['wspd_u'] != 0)                   
-    ds['wspd_x_u'], ds['wspd_y_u'] = calcDirWindSpeeds(ds['wspd_u'], ds['wdir_u']) 
-
-    if ds.attrs['number_of_booms']==2:                                         
-        ds['wdir_l'] = ds['wdir_l'].where(ds['wspd_l'] != 0) 
-        ds['wspd_x_l'], ds['wspd_y_l'] = calcDirWindSpeeds(ds['wspd_l'], ds['wdir_l'])
-
-    if hasattr(ds, 'wdir_i'):
-        if ~ds['wdir_i'].isnull().all() and ~ds['wspd_i'].isnull().all():
-            ds['wdir_i'] = ds['wdir_i'].where(ds['wspd_i'] != 0)                   
-            ds['wspd_x_i'], ds['wspd_y_i'] = calcDirWindSpeeds(ds['wspd_i'], ds['wdir_i'])   
-    
+    if hasattr(ds,'t_i'):
+        if ~ds['t_i'].isnull().all():                                          # Instantaneous msg processing
+            ds['rh_i_cor'] = correctHumidity(ds['rh_i'], ds['t_i'],       # Correct relative humidity
+                                             T_0, T_100, ews, ei0)
 
     ds = clip_values(ds, vars_df)
     return ds
-
-
-def calcDirWindSpeeds(wspd, wdir, deg2rad=np.pi/180):
-    '''Calculate directional wind speed from wind speed and direction
-    
-    Parameters
-    ----------
-    wspd : xr.Dataarray
-        Wind speed data array
-    wdir : xr.Dataarray
-        Wind direction data array
-    deg2rad : float
-        Degree to radians coefficient. The default is np.pi/180
-    
-    Returns
-    -------
-    wspd_x : xr.Dataarray
-        Wind speed in X direction
-    wspd_y : xr.Datarray
-        Wind speed in Y direction
-    '''        
-    wspd_x = wspd * np.sin(wdir * deg2rad)
-    wspd_y = wspd * np.cos(wdir * deg2rad) 
-    return wspd_x, wspd_y
 
 
 def calcCloudCoverage(T, T_0, eps_overcast, eps_clear, dlr, station_id):
