+++ src/pypromice/process/join_l3.py
@@ -4,15 +4,15 @@ import pandas as pd
 import xarray as xr
 from argparse import ArgumentParser
 from pypromice.process import getVars, getMeta, addMeta, getColNames, \
-    roundValues, resampleL2, writeAll
+    roundValues, resampleL3, writeAll
 from pypromice.process.L1toL2 import correctPrecip
 
 def parse_arguments_join():
-    parser = ArgumentParser(description="AWS L2 joiner for merging together two L2 products, for example an L2 RAW and L2 TX data product. An hourly, daily and monthly L2 data product is outputted to the defined output path")
+    parser = ArgumentParser(description="AWS L3 joiner for merging together two L3 products, for example an L3 RAW and L3 TX data product. An hourly, daily and monthly L3 data product is outputted to the defined output path")
     parser.add_argument('-s', '--file1', type=str, required=True,
-                        help='Path to source L2 file, which will be preferenced in merge process')
+                        help='Path to source L3 file, which will be preferenced in merge process')
     parser.add_argument('-t', '--file2', type=str, required=True, 
-                        help='Path to target L2 file, which will be used to fill gaps in merge process')
+                        help='Path to target L3 file, which will be used to fill gaps in merge process')
     parser.add_argument('-o', '--outpath', default=os.getcwd(), type=str, required=True, 
                         help='Path where to write output')
     parser.add_argument('-v', '--variables', default=None, type=str, required=False, 
@@ -41,7 +41,7 @@ def loadArr(infile):
     return ds, name
     
 
-def join_l2():
+def join_l3():
     args = parse_arguments_join()
     
     # Check files
@@ -87,10 +87,10 @@ def join_l2():
         exit()
     
     # Get hourly, daily and monthly datasets
-    print('Resampling L2 data to hourly, daily and monthly resolutions...')
-    l2_h = resampleL2(all_ds, '60min')
-    l2_d = resampleL2(all_ds, '1D')
-    l2_m = resampleL2(all_ds, 'M')
+    print('Resampling L3 data to hourly, daily and monthly resolutions...')
+    l3_h = resampleL3(all_ds, '60min')
+    l3_d = resampleL3(all_ds, '1D')
+    l3_m = resampleL3(all_ds, 'M')
     
     print(f'Adding variable information from {args.variables}...')
         
@@ -98,9 +98,9 @@ def join_l2():
     var = getVars(args.variables)
         	
     # Round all values to specified decimals places
-    l2_h = roundValues(l2_h, var)
-    l2_d = roundValues(l2_d, var)
-    l2_m = roundValues(l2_m, var)
+    l3_h = roundValues(l3_h, var)
+    l3_d = roundValues(l3_d, var)
+    l3_m = roundValues(l3_m, var)
         
     # Get columns to keep
     if hasattr(all_ds, 'p_l'):
@@ -109,22 +109,22 @@ def join_l2():
         col_names = getColNames(var, 1, args.datatype.lower())    
 
     # Assign station id
-    for l in [l2_h, l2_d, l2_m]:
+    for l in [l3_h, l3_d, l3_m]:
         l.attrs['station_id'] = name
     
     # Assign metadata
     print(f'Adding metadata from {args.metadata}...')
     m = getMeta(args.metadata)
-    l2_h = addMeta(l2_h, m)
-    l2_d = addMeta(l2_d, m)
-    l2_m = addMeta(l2_m, m)
+    l3_h = addMeta(l3_h, m)
+    l3_d = addMeta(l3_d, m)
+    l3_m = addMeta(l3_m, m)
       
     # Set up output path
     out = os.path.join(args.outpath, name)
     
     # Write to files
-    writeAll(out, name, l2_h, l2_d, l2_m, col_names)
+    writeAll(out, name, l3_h, l3_d, l3_m, col_names)
     print(f'Files saved to {os.path.join(out, name)}...')
 
 if __name__ == "__main__":  
-    join_l2()
+    join_l3()
